+++ topic12/src/main/java/dao/DaoIn.java
@@ -12,5 +12,5 @@ public interface DaoIn <T extends DaoEntity> {
     List<T> findAll() throws SQLException;
     T findById(int id) throws SQLException;
     Integer insert(T entity) throws SQLException;
-    T remove(Long id);
+    T remove(int id);
 }
